commit dialog

+++ src/app/productos/productos.component.ts
@@ -1,26 +1,37 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatTable } from '@angular/material/table';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
 import { Producto } from '../shared/producto';
 import { ProductoService } from '../shared/producto.service';
+import { ConfirmarComponent } from '../confirmar/confirmar.component';
 
 @Component({
   selector: 'app-productos',
   templateUrl: './productos.component.html',
   styleUrls: ['./productos.component.css']
 })
-export class ProductosComponent implements OnInit {
+export class ProductosComponent implements OnInit, AfterViewInit {
 
   productos : Producto[] = [];
   productoSelected = new Producto();
+
+  dataSource = new MatTableDataSource<Producto>();
   
   mostrarForm = false;
   form = new FormGroup({});
 
-  constructor( private productoService: ProductoService, private formBuilder: FormBuilder ) { }
+  constructor( private productoService: ProductoService, private formBuilder: FormBuilder, public dialog: MatDialog ) { }
 
+  @ViewChild(MatSort) sort!: MatSort;
   //atributo donde almaceno los nombres para las propiedades de las columnas en matTable
-  columnas: string[] = ['id', 'descripcion', 'precio', 'editar', 'borrar']; 
+  columnas: string[] = ['prodId', 'prodDescripcion', 'prodPrecio', 'acciones']; 
+
+  ngAfterViewInit() {
+    this.dataSource.sort = this.sort;
+  }
 
   ngOnInit(): void {
 
@@ -35,9 +46,14 @@ export class ProductosComponent implements OnInit {
     this.productoService.get().subscribe(
       (productos) => {
         this.productos = productos;
-        console.log(productos);
+        this.actualizarTabla();
       } 
     )
+  }
+
+  actualizarTabla() {
+    this.dataSource.data = this.productos;
+    this.dataSource.sort = this.sort;
   }
 
   agregar() {
@@ -48,17 +64,31 @@ export class ProductosComponent implements OnInit {
 
 
   delete(row: Producto) {
-    this.productoService.delete(row.prodId)
-      .subscribe(() => {
-        this.productos = this.productos.filter((x) => {
-          if (x.prodId != row.prodId) {
-            return true
-          } else {
-            return false
-          }
-        });
-      });
-  }
+
+    const dialogRef = this.dialog.open(ConfirmarComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result: ${result}`);
+
+      if (result) {
+        this.productoService.delete(row.prodId)
+          .subscribe(() => {
+
+            //this.items = this.items.filter( x => x !== row);
+
+            this.productos = this.productos.filter((producto) => {
+              if (producto.prodId != row.prodId) {
+                return true
+              } else {
+                return false
+              }
+            });
+
+            this.actualizarTabla();
+          });
+      }
+    });
+}
 
   edit(seleccionado: Producto) {
     this.mostrarForm= true;
@@ -89,6 +119,7 @@ export class ProductosComponent implements OnInit {
         .subscribe((producto) => {
           this.productos.push(producto);
           this.mostrarForm = false;
+          this.actualizarTabla();
         });
 
     }
